% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Check model assumptions}
\usage{
check(x, ...)
}
\arguments{
\item{x}{An object of class "inzts_fit"}

\item{...}{Additional arguments}
}
\value{
The object with the checked and transformed fit
}
\description{
Check model assumptions
}
