% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{decompose}
\alias{decompose}
\title{Decompose a time series object}
\usage{
decompose(obj, multiplicative = FALSE, t = 10, model.lim = NULL,
  data.name = NULL, ...)
}
\arguments{
\item{obj}{an iNZightTS object}

\item{multiplicative}{fit a multiplicative time series model?}

\item{t}{the smoothing parameter}

\item{model.lim}{limits for the time series model}

\item{data.name}{the name of the data}

\item{...}{other, ignored, arguments}
}
\value{
an egg/gtable
}
\description{
Decompose a time series object
}
