% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R, R/forecastplot.R
\name{decomp}
\alias{decomp}
\alias{plot.inz_dcmp}
\alias{plot.inz_frct}
\title{Decompose a time series object}
\usage{
decomp(
  x,
  var = NULL,
  sm_model = c("stl"),
  mult_fit = FALSE,
  model_range = NULL,
  ...
)

\method{plot}{inz_dcmp}(
  x,
  recompose.progress = c(0, 0),
  recompose = any(recompose.progress > 0),
  ylab = NULL,
  xlab = "Date",
  title = NULL,
  xlim = c(NA, NA),
  colour = c("#1B9E46", "#45a8ff", "orangered"),
  ...
)

\method{plot}{inz_frct}(x, xlab = NULL, ylab = NULL, title = NULL, plot = TRUE, ...)
}
\arguments{
\item{x}{an \code{inz_frct} object}

\item{var}{a character vector of length one, or \code{NULL}}

\item{sm_model}{the smoothing method to be used}

\item{mult_fit}{If \code{TRUE}, a multiplicative model is used, otherwise
an additive model is used by default.}

\item{model_range}{range of data to be decomposed by the model, specified as
dates or years, if part of \code{model_range} specified is outside
the range of the data, the exceeding proportion is ignored.}

\item{...}{additional arguments (ignored)}

\item{recompose.progress}{if recompose is \code{TRUE}, this shows how
much to show (for animation!). Length 2 numeric: the first
is 0 for seasonal, and 1 for residual; second component is
how many observations have been recomposed so far}

\item{recompose}{logical as to whether the recomposition is shown or not}

\item{ylab}{a title for the y axis}

\item{xlab}{a title for the x axis}

\item{title}{a title for the graph}

\item{xlim}{the x axis limits}

\item{colour}{vector of three colours for trend, seasonal, and residuals, respectively}

\item{plot}{logical, if \code{FALSE}, the graph isn't drawn}
}
\value{
a decomp (\code{inz_dcmp}) object, a sub-class of dable
}
\description{
Decompose a time series object
}
\examples{
library(dplyr)
d <- visitorsQ \%>\%
    inzightts() \%>\%
    decomp() \%>\%
    plot()

}
\references{
R. B. Cleveland, W. S. Cleveland, J.E. McRae, and I. Terpenning (1990)
STL: A Seasonal-Trend Decomposition Procedure Based on Loess.
Journal of Official Statistics, 6, 3iV73.
}
\seealso{
\code{\link[fabletools]{dable}}
}
