% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fit_model}
\alias{fit_model}
\alias{print.inzts_fit}
\alias{plot.inzts_fit}
\title{Fit model}
\usage{
fit_model(data, x, method = c("arima", "ets"), ...)

\method{print}{inzts_fit}(x, ...)

\method{plot}{inzts_fit}(x, ...)
}
\arguments{
\item{data}{An inz_ts object}

\item{x}{A formula specifying the model}

\item{method}{The method to use for fitting the model}

\item{...}{Additional arguments to pass to the model fitting function}
}
\value{
An object of class "inzts_fit" containing the fitted model
}
\description{
Fit model
}
\section{Functions}{
\itemize{
\item \code{print(inzts_fit)}: Print method

\item \code{plot(inzts_fit)}: Plot method

}}
