% Generated by roxygen2 (4.0.2): do not edit by hand
\name{forecastplot}
\alias{forecastplot}
\title{Forecast plot}
\usage{
forecastplot(vars, ylab = "", multiplicative = FALSE, show = TRUE)
}
\arguments{
\item{vars}{\code{iNZightTS} object}

\item{ylab}{a title for the y axis}

\item{multiplicative}{logical. If \code{TRUE}, a multiplicative model is used,
otherwise an additive model is used by default.}

\item{show}{logical. If \code{TRUE}, draw the plot by default.}
}
\value{
A multiple time series of the predicted values with columns fit,
lwr and upr for the predicted values and the lower and upper bounds
respectively.
}
\description{
Plot a raw time series together with it's fitted curve and add
forecasts and prediction intervals to the end.
}
\details{
The predictions and prediction intervals are the result of models
fitted by the Holt-Winters method. The amount of predicted
observations is calculated by 2 * \code{freq}, where \code{freq} is
the frequency of the time series object.
}
\examples{
z <- iNZightTS(ldeaths)
forecastplot(z)
}
\references{
C.C Holt (1957)
Forecasting seasonals and trends by exponentially weighted
moving averages,
ONR Research Memorandum, Carnigie Institute 52.

P.R Winters (1960)
Forecasting sales by exponentially weighted moving averages,
\emph{Management Science} \bold{6}, 324--342.
}
\seealso{
\code{\link{iNZightTS}},
\code{\link{HoltWinters}}
}

