% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastplot.R
\name{summary.inz_frct}
\alias{summary.inz_frct}
\alias{print.summary_inz_frct}
\title{Summarising forecasts for inzightts objects}
\usage{
\method{summary}{inz_frct}(object, var = NULL, ...)

\method{print}{summary_inz_frct}(x, show_details = FALSE, ...)
}
\arguments{
\item{object}{an \code{inz_frct} object}

\item{var}{a character vector of length one, or \code{NULL}}

\item{...}{additional arguments (ignored)}

\item{x}{a \code{summary_inz_frct} object}

\item{show_details}{logical, if \code{TRUE} the model details will be shown,
only if \code{pred_model = fable::ARIMA} in the \code{predict} call.}
}
\value{
a \code{summary_inz_frct} object, which consists of the first few
        forecast observations, the model used for forecasting and the model
        details (call, coefficients and goodness of fit statistics).
}
\description{
\code{summary} method for objects of class \code{inz_frct}.
}
\examples{
library(dplyr)
s <- visitorsQ \%>\%
    inzightts(var = 2:5) \%>\%
    predict("Japan") \%>\%
    summary("Japan")
s
print(s, show_details = TRUE)

}
\seealso{
\code{\link[iNZightTS]{predict.inz_ts}}
}
