% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplot.R
\name{ggplotable}
\alias{ggplotable}
\title{Preliminary check for a plotly::ggplotly() call}
\usage{
ggplotable(x)
}
\arguments{
\item{x}{a \code{ggplot} object produced by iNZightTS}
}
\value{
a \code{logical}
}
\description{
Check if a plot generated by iNZightTS can be passed to plotly::ggplotly().
}
\examples{
x <- inzightts(visitorsQ)
\dontrun{
ggplotable(plot(x))
ggplotable(plot(x, names(x)[-1]))
}

}
\seealso{
\code{\link[plotly]{ggplotly}}
}
