% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightTS2-package.R
\docType{package}
\name{iNZightTS2-package}
\alias{iNZightTS2-package}
\title{Time Series Data Analysis}
\description{
The iNZightTS2 package provides some simple analysis tools for
exploring time series data. It is used in the iNZight software.
}
\seealso{
\code{\link[iNZightTS2]{inzightts}}
}
\author{
Tom Elliott (previously: Marco Kuper, Simon Potter, and David Banks)
}
\keyword{iNZight}
