% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastplot.R
\name{log_if}
\alias{log_if}
\title{Conditional log transform}
\usage{
log_if(x, mult_fit)
}
\arguments{
\item{x}{A numeric vector to be transformed.}

\item{mult_fit}{Logical; set to \code{TRUE} for a multiplicative model, or \code{FALSE} for the additive model.}
}
\value{
A numeric vector.
}
\description{
Conditional log transform
}
