% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareplot.R
\name{plot.iNZightMTS}
\alias{plot.iNZightMTS}
\title{Plot multiple time series}
\usage{
\method{plot}{iNZightMTS}(x, ...)
}
\arguments{
\item{x}{Multiple time series object}

\item{...}{additional arguments}

\item{vars}{iNZightTS object with the data and frequency information included}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{multiplicative}{logical, if TRUE multiplcative series will be used}

\item{t}{smoothing parameter}
}
\description{
Plot a multiple time series object to compare several series
}
\author{
iNZight
}
