% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplot.R
\name{pred}
\alias{pred}
\title{Get forecast prediction values}
\usage{
pred(x)
}
\arguments{
\item{x}{the forecast object (a plot with predictions)}
}
\value{
a time series forecasts object
}
\description{
Get forecast prediction values
}
