% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.iNZightTS}
\alias{print.iNZightTS}
\title{Print an iNZightTS object}
\usage{
\method{print}{iNZightTS}(x, full = FALSE, ...)
}
\arguments{
\item{x}{the \code{iNZightTS} object to be printed}

\item{full}{whether to print all the underlying data}

\item{...}{Unused arguments. Only here for consistency with the base S3 method.}
}
\description{
Print method for iNZightTS (time series) objects.
}
\details{
The \code{full} argument controls whether to print all the data from
which the \code{iNZightTS} object has been created. The default is set
to \code{FALSE} and only the \code{head()} of the data will be printed.
}
\examples{
print(iNZightTS(UKgas))
}
\seealso{
\code{\link{print}}, \code{\link{iNZightTS}}
}

