% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{seaice}
\alias{seaice}
\title{Sea Ice}
\format{
A data frame with 265 rows and 3 variables:
\describe{
  \item{Time}{The time variable}
  \item{Arctic}{Sea ice measurement for the Arctic}
  \item{Antarctica}{Sea ice measurement for Antarctica}
}
}
\usage{
seaice
}
\description{
A dataset containing sea ice measurements from 1990 to 2011.
}
\keyword{datasets}
