% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.R
\name{subseries}
\alias{subseries}
\title{Seasonal subseries plots for inzightts}
\usage{
subseries(
  x,
  var = NULL,
  show_mean = TRUE,
  xlab = NULL,
  ylab = NULL,
  title = NULL
)
}
\arguments{
\item{x}{an inzightts (\code{inz_ts}) or tsibble (\code{tbl_ts}) object}

\item{var}{a character vector of the variable(s) to be plotted, or \code{NULL}}

\item{show_mean}{logical, if \code{FALSE}, the mean line isn't drawn}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{title}{a title for the graph}
}
\value{
a \code{numeric}
}
\description{
A time series plot faceted by seasonal period.
}
\examples{
t <- inzightts(visitorsQ)
\dontrun{
subseries(t)
}

}
\seealso{
\code{\link[feasts]{gg_subseries}}
}
