% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{visitorsM2}
\alias{visitorsM2}
\title{Visitors (monthly)}
\format{
A data frame with 164 rows and 5 variables:
\describe{
  \item{Time}{The time variable (year/month)}
  \item{Australia}{Visitor counts for Australia}
  \item{China..People.s.Republic.of}{Visitor counts for China}
  \item{Japan}{Visitor counts for Japan}
  \item{United.Kingdom}{Visitor counts for the UK}
}
}
\usage{
visitorsM2
}
\description{
A dataset containing monthly visitor numbers for several countries.
}
\keyword{datasets}
