% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareplot.R
\name{compareplot}
\alias{compareplot}
\alias{compareplot.1}
\alias{compareplot.2p}
\title{Comparison plot - depreciated}
\usage{
compareplot(x, ...)

compareplot.1(vars, xlab = "Time", ylab = "", multiplicative = FALSE,
  t = 0)

compareplot.2p(vars, xlab = "Time", ylab = "", multiplicative = FALSE,
  t = 0)
}
\description{
Comparison plot - depreciated
}
\section{Functions}{
\itemize{
\item \code{compareplot.1}: For series with frequency = 1 (DEPRECIATED)

\item \code{compareplot.2p}: For series with frequency > 1
}}

