% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareplot.R
\name{compareplot}
\alias{compareplot}
\title{Comparison plot - depreciated}
\usage{
compareplot(x, ...)
}
\arguments{
\item{x}{an iNZightTS object}

\item{...}{additional arguments passed to `plot()`}
}
\value{
No return value, called for the side effect of drawing a plot.
}
\description{
Comparison plot - depreciated
}
