% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastplot.R
\name{forecastplot}
\alias{forecastplot}
\title{Forecast plot - DEPRECATED}
\usage{
forecastplot(x, ...)
}
\arguments{
\item{x}{\code{iNZightTS} object}

\item{...}{additional arguments passed on}
}
\value{
Called for the side effect of drawing a plot.
        The constructed \code{ggplot} object is returned invisibly.
}
\description{
Plot a raw time series together with it's fitted curve and add
forecasts and prediction intervals to the end.
}
\details{
The predictions and prediction intervals are the result of models
fitted by the Holt-Winters method. The amount of predicted
observations is calculated by 2 * \code{freq}, where \code{freq} is
the frequency of the time series object.
}
