\name{iNZightTS}
\docType{package}
\title{
  Time Series Data Analysis
}
\description{
  The iNZightTS package provides some simple analysis tools for
  exploring time series data. It is used in the iNZight software.
}
\details{
\tabular{ll}{
  Package: \tab iNZightTS\cr
  Type: \tab Package\cr
  Version: \tab 1.0\cr
  Date: \tab 2014-02-25\cr
  License: \tab GPL-2\cr
}
}
\author{
  Marco Kuper, originally developed by Simon Potter and David Banks.
  
  Maintainer: Marco Kuper <MarcoKuper42@gmail.com>
}
\keyword{ iNZight }
\seealso{
  \code{\link{iNZight}}
}
\examples{

}
