% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightTS-package.R
\docType{package}
\name{iNZightTS-package}
\alias{iNZightTS-package}
\title{Time Series Data Analysis}
\description{
The iNZightTS package provides some simple analysis tools for
exploring time series data. It is used in the iNZight software.
}
\seealso{
\code{iNZight}
}
\author{
Tom Elliott (previously: Marco Kuper, Simon Potter, and David Banks)
}
\keyword{iNZight}
