% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightTS.R
\name{inzightts}
\alias{inzightts}
\alias{inzightts.character}
\alias{inzightts.data.frame}
\alias{inzightts.ts}
\alias{inzightts.tbl_ts}
\title{Coerce to an inzightts (Time-Series) Objects}
\usage{
inzightts(x, ...)

\method{inzightts}{character}(x, stringsAsFactors = TRUE, as.is = TRUE, ...)

\method{inzightts}{data.frame}(
  x,
  var = NULL,
  index = NULL,
  key = NULL,
  start = NULL,
  end = NULL,
  freq = NULL,
  ...
)

\method{inzightts}{ts}(x, var_name = NULL, pivot_longer = FALSE, ...)

\method{inzightts}{tbl_ts}(x, ...)
}
\arguments{
\item{x}{a \code{data.frame}, \code{ts}, tsibble, or path}

\item{...}{additional arguments to be passed to or from methods}

\item{stringsAsFactors}{see \code{\link[utils]{read.csv}}}

\item{as.is}{see \code{\link[utils]{read.csv}}}

\item{var}{the column number or name for the observations used
from \code{data} in the actual time series}

\item{index}{which column contains the time variable}

\item{key}{Variable(s) that uniquely determine time indices}

\item{start}{the time of the first observation.
Either a single number or a vector
of two integers, which specify a natural time unit
and a (1-based) number of samples into the time unit}

\item{end}{the time of the last observation, specified in the
same way as \code{start}}

\item{freq}{the number of observations per unit of time}

\item{var_name}{rename the variable column of the univariate time series,
applicable only if \code{x} is not an \code{mts} object.}

\item{pivot_longer}{logical, \code{TRUE} gives a "longer" form of the data,
otherwise as is, applicable only if \code{x} is an \code{mts} object.}
}
\value{
an inzightts (\code{inz_ts}) object, a sub-class of tsibble
        which includes the index variable, temporal variable and,
        if applicable, relevant keys.
}
\description{
The function \code{inzightts} is used to create temporal data frames
used in iNZight.
}
\details{
The function \code{inzightts} is used to create temporal data
frames. Unlike \code{ts} objects, these are tsibble objects
that enables temporal data wrangling adapting to the tidy
data principles, which are both data- and model-oriented.

If a \code{ts} object is used to create the inzightts object,
all the domain information is extracted from that object.

\code{index} should be a \code{character}, \code{Date},
\code{\link[tsibble]{yearweek}}, \code{\link[tsibble]{yearmonth}} or
\code{\link[tsibble]{yearquarter}} object.

If \code{index} is a \code{character}, the function recognises the
following time variable formatS without case sensitive:
 \itemize{
  \item "(Y)yyyy" annually data e.g."(Y)1991"
  \item "(Y)yyyyMmm" monthly data e.g."(Y)1991M01"
  \item "(Y)yyyyQqq" quarterly data e.g."(Y)1991Q01"
  \item "(Y)yyyyWww" weekly data with yearly seasonality e.g."(Y)1991W01"
  \item "(Y)yyyyDdd" daily data with yearly seasonality e.g."(Y)1991D01"
  \item "WwwDdd"  daily data with weekly seasonality e.g. "W01D01"
  \item "DddHhh" hourly data with daily seasonality e.g. "D01H01"
}
The length of digits of each time unit could be flexible and allowing space
between the time unit

In case of \code{data} being a data.frame or path to a \code{.csv}
file and \code{start} being omitted, the starting date and the
\code{freq} is extracted from the column that includes the time
information. This column is either named \code{"Time"} or is the first
column. If \code{end} is omitted, all of the data will be used for the
time-series.
}
\examples{
# create from a ts object
z <- inzightts(UKgas)
\dontrun{
plot(z)
}

# create from a data.frame
x <- inzightts(
    data.frame(Return = rnorm(100), Time = 1900:1999),
    var = "Return"
)
# or specify a time column
x <- inzightts(
    data.frame(Return = rnorm(100), Year = 1900:1999),
    var = "Return", index = "Year"
)

# create from a data.frame with modified time frame
y <- inzightts(
    data.frame(Return = rnorm(100)),
    start = c(1990, 1), end = c(1993, 5), freq = 12, var = 1
)
\dontrun{
plot(y)
}

}
\seealso{
\code{\link[tsibble]{tsibble}}, \code{\link[tsibble]{as_tsibble}}
         and \code{\link[tsibble]{new_tsibble}}
}
