% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightTS.R
\name{inzightts}
\alias{inzightts}
\title{inzightts (Time-Series) Objects}
\usage{
inzightts(
  data,
  start = 1,
  end,
  freq = 1,
  var = 2,
  time.col = grep("time", names(data), ignore.case = TRUE)[1],
  ...
)
}
\arguments{
\item{data}{a \code{data.frame} containing time information and
observation or a path to a \code{.csv} file with
such information or a \code{ts} object}

\item{start}{the time of the first observation.
Either a single number or a vector
of two integers, which specify a natural time unit
and a (1-based) number of samples into the time unit}

\item{end}{the time of the last observation, specified in the
same way as \code{start}}

\item{freq}{the number of observations per unit of time}

\item{var}{the column number or name for the observations used
from \code{data} in the actual time series}

\item{time.col}{which column contains the time variable}

\item{...}{additional information passed to \code{read.csv()} and used when
\code{data} is a path}

\item{ignore.case}{logical, ignore the case?}
}
\value{
an inzightts (\code{inz_ts}) object, a sub-class of tsibble
        which includes the index variable, temporal variable and,
        if applicable, relevant keys.
}
\description{
The function \code{inzightts} is used to create temporal data frames
used in iNZight.
}
\details{
The function \code{inzightts} is used to create temporal data
frames. Unlike \code{ts} objects, these are tsibble objects
that enables temporal data wrangling adapting to the tidy
data principles, which are both data- and model-oriented.

If a \code{ts} object is used to create the inzightts object,
all the domain information is extracted from that object.

The function recognises the following time variable formatS without case sensitive:
 \itemize{
  \item "(Y)yyyy" annually data e.g."(Y)1991"
  \item "(Y)yyyyMmm" monthly data e.g."(Y)1991M01"
  \item "(Y)yyyyQqq" quarterly data e.g."(Y)1991Q01"
  \item "(Y)yyyyWww" weekly data with yearly seasonality e.g."(Y)1991W01"
  \item "(Y)yyyyDdd" daily data with yearly seasonality e.g."(Y)1991D01"
  \item "WwwDdd"  daily data with weekly seasonality e.g. "W01D01"
  \item "DddHhh" hourly data with daily seasonality e.g. "D01H01"
}
The length of digits of each time unit could be flexible and allowing space between the
time unit

In case of \code{data} being a data.frame or path to a \code{.csv}
file and \code{start} being omitted, the starting date and the
\code{freq} is extracted from the column that includes the time
information. This column is either named \code{"Time"} or is the first
column. If \code{end} is omitted, all of the data will be used for the
time-series.
}
\examples{
# create from a ts object
z <- inzightts(UKgas)
plot(z)

# create from a data.frame
x <- inzightts(data.frame(Return = rnorm(100), Time = 1900:1999),
    var = "Return")
# or specify a time column
x <- inzightts(data.frame(Return = rnorm(100), Year = 1900:1999),
    var = "Return", time.col = "Year")

# create from a data.frame with modified time frame
y <- inzightts(data.frame(Return = rnorm(100)),
    start = c(1990, 1), end = c(1993, 5), freq = 12, var = 1)
plot(y)

}
\seealso{
\code{\link[tsibble]{tsibble}}
}
