% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastplot.R
\name{log_if}
\alias{log_if}
\title{Toggleable log logarithmic transformation}
\usage{
log_if(x, mult_fit)
}
\arguments{
\item{x}{a \code{numeric}}

\item{mult_fit}{whether or not to apply logarithmic transformation to the input}
}
\value{
a \code{numeric}
}
\description{
Log-transform the input if \code{mult_fit = TRUE}, else return the original input.
}
\examples{
x <- runif(1e4, 1, 100)
all.equal(log_if(x, TRUE), log(x))
all.equal(log_if(x, FALSE), x)

}
\seealso{
\code{\link[fabletools]{new_transformation}}
}
