% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiseries.R
\name{multiseries}
\alias{multiseries}
\title{Compare multiple timeseries - DEPRECIATED}
\usage{
multiseries(x, ...)
}
\arguments{
\item{x}{iNZightMTS object containing data}

\item{...}{Further arguments to be passed to `plot()`}
}
\description{
Compare multiple timeseries - DEPRECIATED
}
