% Generated by roxygen2 (4.0.2): do not edit by hand
\name{multiseries}
\alias{multiseries}
\title{Compare multiple timeseries}
\usage{
multiseries(x, ...)
}
\arguments{
\item{x}{iNZightMTS object containing data}

\item{...}{Further arguments to be passed onto specific methods.}
}
\description{
Draw a plot to compare 2 or more timeseries with each
other. The resulting plot contains the original timeseries as well as
plots showing the season effects of each timeseries, if the frequency
is greater than 1.
}
\details{
\code{x} is an \code{\link{iNZightMTS}} object containing some
data for which timeseries objects can be created. The data columns
used for the plotting are those that are used in the \code{\link{iNZightMTS}}
object.
\cr \cr
The frequency used for the plotting is also stored in \code{x}.
}
\examples{
# Frequency > 1
y <- iNZightTS(Seatbelts)
multiseries(y)

# Frequency = 1
# Casulties for Jan of each year
X <- Seatbelts[12*(1:15), -8]
X <- as.ts(X)
z <- iNZightTS(X)
multiseries(z)
}
\seealso{
\code{\link{iNZightTS}}
}
\keyword{timeseries}

