% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplot.R
\name{rawplot}
\alias{rawplot}
\title{Draw a simple time series plot}
\usage{
rawplot(obj, multiplicative = FALSE, ylab = "", xlab = "",
  animate = FALSE, t = 0, e = NULL)
}
\arguments{
\item{obj}{an \code{iNZightTS} object}

\item{multiplicative}{logical. If \code{TRUE}, a multiplicative model is used,
otherwise an additive model is used by default.}

\item{ylab}{a title for the y axis}

\item{xlab}{a title for the x axis}

\item{animate}{animate the plotting process?}

\item{t}{smoothing parameter}

\item{e}{\code{NULL} by default to support animation stop}
}
\description{
Draws a plot of a given \code{iNZightTS} object with the trend superimposed.
}
\details{
If animate is set to \code{TRUE}, a scatterplot of all points in the
time series will appear followed by slowly drawn lines connecting the
points, simulating the drawing of a time series by hand.
}
\keyword{timeseries}
