% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplot.R
\name{plot.inz_ts}
\alias{plot.inz_ts}
\title{Draw a simple time series plot}
\usage{
\method{plot}{inz_ts}(
  x,
  var = NULL,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  xlim = NULL,
  aspect = NULL,
  compare = TRUE,
  pal = NULL,
  smoother = TRUE,
  sm_model = "stl",
  mult_fit = FALSE,
  emphasise = NULL,
  non_emph_opacity = 0.2,
  show_iso_obs = TRUE,
  iso_obs_size = 1,
  ...
)
}
\arguments{
\item{x}{an inzightts (\code{inz_ts}) object}

\item{var}{a character vector of the variable(s) to be plotted, or \code{NULL}}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{title}{a title for the graph}

\item{xlim}{axis limits, specified as dates or years}

\item{aspect}{the aspect ratio of the plot;
it will be about \code{aspect} times wider than it is high}

\item{compare}{logical, whether to plot the key levels in a single plot}

\item{pal}{(only if a categorical variable is passed to \code{var}): the
colour palette for the categorical plot, the palette vector should
be in the same order per the rows of \code{tsibble::key_data(x)}.}

\item{smoother}{logical, if \code{TRUE} the smoother will be drawn}

\item{sm_model}{the smoothing method to be used}

\item{mult_fit}{If \code{TRUE}, a multiplicative model is used, otherwise
an additive model is used by default.}

\item{emphasise}{integer vector to specify the key level(s) to focus in the
plot. The integer maps to the specific key level(s) corresponding to
the ith row of \code{tsibble::key_data(x)}.}

\item{non_emph_opacity}{numeric. If \code{(0, 1]}, this argument determines
the opacity of the series other than the focused one(s) (to highlight
the focused series). If \code{non_emph_opacity = 0}, the plot draws
the focused series in its own scales.}

\item{show_iso_obs}{logical, whether to plot isolated observations between
time series gaps or not, if there is any}

\item{iso_obs_size}{numeric, scaling the size of isolated observations, if
\code{show_iso_obs = TRUE} and they exists}

\item{...}{additional arguments (ignored)}
}
\value{
a time series plot (constructed with ggplot2) is returned, which
        can be added to if desired.
}
\description{
Draws a plot of a given inzightts (\code{inz_ts}) object with the trend superimposed.
}
\examples{
t <- inzightts(visitorsQ, var = c(2, 4))

\dontrun{
plot(t)
plot(t, var = names(t)[-1])
plot(t, var = "Japan")
plot(t, mult_fit = TRUE)
}

}
\seealso{
\code{\link[tsibble]{key_data}}
}
\keyword{timeseries}
