% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplot.R
\name{plot.inz_ts}
\alias{plot.inz_ts}
\title{Draw a simple time series plot}
\usage{
\method{plot}{inz_ts}(
  x,
  var = NULL,
  xlab = NULL,
  ylab = NULL,
  title = NULL,
  plot = TRUE,
  xlim = NULL,
  aspect = NULL,
  compare = FALSE,
  smoother = TRUE,
  sm_model = "stl",
  mult_fit = FALSE,
  ...
)
}
\arguments{
\item{x}{an inzightts (\code{inz_ts}) object}

\item{var}{a character vector of the variable(s) to be plotted, or \code{NULL}}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{title}{a title for the graph}

\item{plot}{logical, if \code{FALSE}, the graph isn't drawn}

\item{xlim}{axis limits, specified as dates or years}

\item{aspect}{the aspect ratio of the plot;
it will be about \code{aspect} times wider than it is high}

\item{compare}{logical (not used)}

\item{smoother}{logical, if \code{TRUE} the smoother will be drawn}

\item{sm_model}{the smoothing method to be used}

\item{mult_fit}{If \code{TRUE}, a multiplicative model is used, otherwise
an additive model is used by default.}

\item{...}{additional arguments (ignored)}
}
\value{
a time series plot (constructed with ggplot2) is returned invisibly,
        which can be added to if desired.
}
\description{
Draws a plot of a given inzightts (\code{inz_ts}) object with the trend superimposed.
}
\examples{
t <- inzightts(visitorsQ, var = c(2, 4))
plot(t)
plot(t, var = names(t)[-1])
plot(t, var = "Japan")
plot(t, mult_fit = TRUE)

}
\keyword{timeseries}
