% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawplot.R
\name{rawplot}
\alias{rawplot}
\title{Time series plot - depreciated}
\usage{
rawplot(...)
}
\arguments{
\item{...}{arguments passed to `plot` method}
}
\value{
Called to draw a plot. Inivisbly returns a \code{ggplot} object.
}
\description{
Time series plot - depreciated
}
