% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recompose.R
\name{recompose}
\alias{recompose}
\title{Recompose a decomposed time series}
\usage{
recompose(...)
}
\arguments{
\item{...}{additional arguments, ignored}
}
\value{
the recomposed series
}
\description{
Recompose a time series object, with optional animation.
}
\author{
iNZight
}
