% Generated by roxygen2 (4.0.2): do not edit by hand
\name{seasonplot}
\alias{seasonplot}
\title{Plot Seasonal Subseries from a Time Series}
\usage{
seasonplot(obj, ...)
}
\arguments{
\item{obj}{an \code{iNZightTS} object}

\item{...}{Further arguments to be passed onto specific methods.}
}
\description{
This function plots the seasonal components of a time series together
with the estimated seasonal effects of that series.
}
\details{
The resulting window will contain two plots. On the left, every
seasonal subseries of the time series is plotted. On the right will be
the average seasonal effect of the series.
}
\examples{
x <- iNZightTS(UKgas)
seasonplot(x)
}
\seealso{
\code{\link{iNZightTS}}
}

